requirements.txt
Flask
Flask-RESTful
Flask-Bootstrap

Jinja2
Werkzeug

itsdangerous
gunicorn

async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachelib==0.9.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
dominate==2.7.0
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-OAuthlib==0.9.6
Flask-Session==0.5.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==2.1.0
outcome==1.2.0
packaging==23.1
pandas==1.5.3
plotly==5.15.0
psycopg2-binary==2.9.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.1.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stripe==5.5.0
tenacity==8.2.2
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
user-agent==0.1.10
visitor==0.1.3
webdriver-manager==4.0.0
Werkzeug==2.3.6
whichcraft==0.6.1
wsproto==1.2.0

# json
# importlib

# to test mysql connectivity
mysql-connector

## is json native? not able to pip install and seems to import without installing
## is importlib native? not able to pip install and seems to import without installing
